api: Refactor into a folder
